fixes in line with comments on PR
