+++ src/firebase.ts
@@ -25,5 +25,3 @@ const app = initializeApp(firebaseConfig);
 export const auth = getAuth(app);
 
 export const db = getFirestore(app);
-
-export default getFirestore(app);
